PreferenceManager pt. 1

+++ Implementation/ApplicationServer/src/main/java/it/polimi/travlendarplus/entities/Location.java
@@ -33,7 +33,7 @@ public class Location extends GenericEntity {
     }
 
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -41,7 +41,7 @@ public class Location extends GenericEntity {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
